# Speeding up strain rate and deviatoric stress kernels